# docs/05-api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Klick-Pro (servio) HTTP API
  version: 0.0.0-reverse-engineered
  description: >-
    Reverse-engineered from source at commit cd8f4fb (verified 2026-09-16). Source of truth is the implementation; every
    operation carries x-source file:line.


    Conventions and caveats:

    - Responses are raw NextResponse.json; errors are mostly {"error": string}. The apiSuccess/apiError envelope in
    src/lib/api-response.ts is not used.

    - Every POST/PUT/PATCH/DELETE under /api/* must carry an Origin header equal to the app origin or APP_URL, otherwise
    proxy.ts returns 403 {"error":"Request origin is not allowed."}. This applies to webhooks and non-browser clients
    too. Under the custom server only the exact APP_URL origin passes in practice; a genuine same-origin request to a
    different host (e.g. 127.0.0.1 when APP_URL is localhost) also gets 403. [VALIDATED 2026-09-17 · V-20]

    - /api/v1/:path* is rewritten to /api/:path* (next.config.ts, afterFiles). Physical routes /api/v1/messages and
    /api/v1/professionals take precedence and have no /api/* twin. /api/v1/v1/* is not rewritten twice and returns
    404. [VALIDATED 2026-09-17 · V-13]

    - Schemas marked [NEEDS VALIDATION] could not be fully established from code; no properties were invented.

    - Realtime Socket.IO events (path /api/realtime) are documented in docs/05-api/api-specification.md, not here.


    Human-readable reference: docs/05-api/api-specification.md
servers:
...
paths:
...
  /api/wallet/deposit/fail:
    post:
      tags:
        - Wallet
      summary: Mark the caller's pending top-up FAILED (checkout cancelled)
      operationId: failWalletDeposit
      security:
        - cookieAuth: []
      description: Source app/api/wallet/deposit/fail/route.ts:11
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orderId
              properties:
                orderId:
                  type: string
                  minLength: 1
                reason:
                  type: string
                  maxLength: 240
      responses:
        "200":
          description: OK (no-op if already completed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
        "400":
          description: Invalid payment order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Wallet payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/wallet/milestone:
    post:
      tags:
        - Wallet
      summary: Fund an AWAITING_CLIENT_REVIEW milestone from the client wallet (escrow to admin)
      operationId: fundMilestoneFromWallet
      security:
        - cookieAuth: []
      description: >-
        Source app/api/wallet/milestone/route.ts:14. One transaction: claim milestone, upsert Payment, debit client and
        credit admin wallet ledger, Payment FUNDED, upsert Invoice, milestone AWAITING_ADMIN_APPROVAL,
        ProjectTransaction WALLET_MILESTONE_FUNDED. Then notifies the professional and admins and emits project:updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - projectId
                - milestoneId
              properties:
                projectId:
                  type: integer
                  minimum: 1
                milestoneId:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: Funded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestonePaymentResult"
        "400":
          description: Invalid milestone payment request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Client sign-in is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: Insufficient wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Client access is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Not ready for payment or already paid/being paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Milestone payment could not be completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: servio_session
      description: >-
        HS256 JWT {userId, role, sessionId} issued by src/lib/auth.ts createSession; validated against the sessions
        table on every request. Same cookie for clients, professionals and admins — role is inside the token.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Same session JWT sent as Authorization: Bearer. Accepted only by the few handlers noted per operation."
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    OkTrue:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          const: true
...
    MilestonePaymentResult:
      type: object
      description: >-
        Returned by POST /api/wallet/milestone (status FUNDED) and project-actions approve-milestone on OFFLINE jobs
        (status COMPLETED)
      properties:
        ok:
          const: true
        paymentMethod:
          const: OFFLINE
          description: Only for the offline variant
        charged:
          type: integer
        professionalReceives:
          type: integer
        adminReceives:
          type: integer
          description: "Wallet variant: equals charged; offline: 0"
        platformEarnings:
          type: integer
        remainingBalance:
          type: integer
          description: Wallet variant only
        status:
          type: string
          enum:
            - FUNDED
            - COMPLETED
        message:
          type: string
...